Move Log middleware to log_middleware.go
